feat: dev change password
