+++ src/Template/Pages/Accounts/Password/Change/index.js
@@ -1,7 +1,7 @@
 import React from "react";
 import BasicTemplate from "Template/BasicTemplate";
 import WrapperMenuAccounts from "Template/Pages/Accounts/WrapperMenuAccount";
-import ChangePassword from "./ChangePassword";
+import ChangePassword from "../Update";
 
 const Change = () => {
   return (
