+++ src/Template/Pages/Accounts/Password/Forgot/MouseParallax.js
@@ -10,19 +10,19 @@ const MouseParallax = ({ mouseParallax }) => {
   return (
     <div className="icon__wrapper">
       <animated.div
-        class="card1"
+        className="card1"
         style={{ transform: mouseParallax.xy.interpolate(trans1) }}
       />
       <animated.div
-        class="card2"
+        className="card2"
         style={{ transform: mouseParallax.xy.interpolate(trans2) }}
       />
       <animated.div
-        class="card3"
+        className="card3"
         style={{ transform: mouseParallax.xy.interpolate(trans3) }}
       />
       <animated.div
-        class="card4"
+        className="card4"
         style={{ transform: mouseParallax.xy.interpolate(trans4) }}
       />
     </div>
