+++ src/Template/Pages/Accounts/Password/Forgot/index.js
@@ -1,0 +1,10 @@
+import React from "react";
+// import BasicTemplate from "Template/BasicTemplate";
+import ResetPassword from "./ForgotPassword";
+import "./forgot.scss";
+
+const Reset = () => {
+  return <ResetPassword />;
+};
+
+export default Reset;
