+++ src/Template/Pages/Accounts/Password/Reset/reset.scss
@@ -1,144 +1,18 @@
-.RP {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
+.reset-password {
+  position: relative;
+  display: flex;
+  align-items: center;
 
-    &__content {
-        background-color: rgba(var(--cdc, 255, 255, 255), 1);
-        border: 1px solid rgba(var(--b6a, 219, 219, 219), 1);
-        border-radius: 3px;
-        margin: 60px auto 0;
-        max-width: 388px;
-        overflow: hidden;
-
-        @media (max-width:388px) {
-            padding: 0 16px;
-            border: 0;
-            margin: 0;
-        }
-
-        &--item {
-            margin-bottom: 16px;
-            padding: 0 44px;
-
-            @media (max-width:388px) {
-                padding: 0 16px;
-            }
-
-            &:last-child {
-                padding: 0;
-                margin-top: 68px;
-                margin-bottom: 0;
-            }
-        }
-
-        &--icon {
-            margin-top: 32px;
-            display: flex;
-            justify-content: center;
-
-            .icon__wrapper {
-                width: 100%;
-                height: 100%;
-                display: flex;
-                align-items: center;
-                justify-content: center;
-            }
-
-            // .card1,
-            .card2,
-            .card3,
-            .card4 {
-                position: absolute;
-                border-radius: 5px;
-                background-size: cover;
-                background-position: center center;
-                background-repeat: no-repeat;
-                will-change: transform;
-            }
-
-            .card1 {
-                min-width: 60px;
-                min-height: 60px;
-                width: 45vw;
-                height: 45vw;
-                max-width: 100px;
-                max-height: 100px;
-                background-image: url(https://image.flaticon.com/icons/svg/119/119596.svg);
-            }
-
-            .card2 {
-                width: 25px;
-                height: 25px;
-                background-image: url(https://image.flaticon.com/icons/svg/789/789395.svg);
-            }
-
-            .card3 {
-                opacity: 0.9;
-                width: 25px;
-                height: 25px;
-                background-image: url(https://image.flaticon.com/icons/svg/414/414927.svg);
-            }
-
-            .card4 {
-                width: 25px;
-                height: 25px;
-                background-image: url(https://image.flaticon.com/icons/svg/789/789392.svg);
-            }
-        }
-
-        .item {
-            &__TLI {
-                font-weight: 600;
-                font-size: 16px;
-                line-height: 24px;
-            }
-
-            &__description {
-                color: rgba(var(--f52, 153, 153, 153), 1);
-                text-align: center;
-                display: block;
-                font-size: 14px;
-                line-height: 18px;
-                font-weight: 400;
-            }
-
-            &__devider {
-                color: #999;
-                font-size: 13px;
-                font-weight: 600;
-                line-height: 15px;
-            }
-
-            &__CNA {
-                color: rgba(var(--f07, 38, 38, 38), 1);
-                border: 0;
-                background-color: transparent;
-                text-decoration: none;
-                font-weight: 600;
-                font-size: 14px;
-                line-height: 18px;
-            }
-
-            &__BTL {
-                height: 44px;
-                border: 1px solid rgba(var(--b38, 219, 219, 219), 1);
-                display: flex;
-                align-items: center;
-                justify-content: center;
-
-                &--content {
-                    color: rgba(var(--f07, 38, 38, 38), 1);
-                    background-color: transparent;
-                    text-decoration: none;
-                    font-weight: 600;
-                    font-size: 14px;
-                    line-height: 18px;
-                    width: 100%;
-                }
-            }
-        }
-    }
-
+  &__content {
+    max-width: 388px;
+    background-color: rgba(var(--cdc, 255, 255, 255), 1);
+    border: 1px solid rgba(var(--b6a, 219, 219, 219), 1);
+    border-radius: 3px;
+    margin-right: auto;
+    margin-left: auto;
+    min-height: 0;
+    min-width: 0;
+    box-sizing: border-box;
+    padding: 24px;
+  }
 }
