+++ src/Template/pageConfigs.js
@@ -3,6 +3,7 @@ import Login from "./Pages/Accounts/Login";
 import EmailSignup from "./Pages/Accounts/EmailSignup";
 import PersonalPost from "./Pages/PersonalPage/PersonalPosts";
 import PersonalSaved from "./Pages/PersonalPage/PersonalSaved";
+import ForgotPassword from "Template/Pages/Accounts/Password/Forgot";
 import ResetPassword from "Template/Pages/Accounts/Password/Reset";
 import EditProfile from "Template/Pages/Accounts/EditProfile";
 import ChangePassword from "Template/Pages/Accounts/Password/Change";
@@ -41,6 +42,11 @@ const pageConfigs = [
   },
   {
     path: "/accounts/password/reset/",
+    component: ForgotPassword,
+    exact: true
+  },
+  {
+    path: "/accounts/password/reset/:token",
     component: ResetPassword,
     exact: true
   },
